Fixes links contents removed by Grav links post-processing
There were a but in the DOM-parsing lib we're using to post-process Grav
links and images, so I implemented a workaround, but it stripped all
HTML inside links (so inline codes or emphasis were not possible in
links). The lib implementation now works correctly so we switched to
it and everything work now.

Also we don't process footnotes backlinks as it's useless.